backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
openai==0.27.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-jose==3.3.0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
yarl==1.15.0